template: only use Go font for "go-fonts" post
This change reverts the blog to use the system default font for all
pages except the go-fonts post.

Style selectors have been added to select that post and apply the fonts
to it both on its own page as well as on the homepage.

Regardless of anyone's personal opinions about the font, there were
several negative comments about the font on social media sites for posts
unrelated to the font. We want to encourage people to discuss the content
of our articles rather than get tied up about their presentation.

Change-Id: I107f57e289f408584e2e5f6a8de3a6a445ea42a4
Reviewed-on: https://go-review.googlesource.com/37137
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>