README, template: add more links
Move the README to README.md so Gerrit can render it; currently
Gerrit only renders files named exactly "README.md" (for example at
https://go.googlesource.com/go).

Add more links to the README explaining how to file issues,
how to submit code changes, where to download the code to and
how to get it. Hopefully this should help people who go to
https://go.googlesource.com/blog or https://github.com/golang/blog
figure out how to get started with development.

Add a link in the footer to the source code for the blog, so people
know where to find it.

Updates golang/go#21038.
Updates golang/go#21034.

Change-Id: Ie378cf2aa42f56485b5d5c9a943f5b9957d96371
Reviewed-on: https://go-review.googlesource.com/49291
Reviewed-by: Russ Cox <rsc@golang.org>