blog: user- and test-friendly local server
Log a short message when the blog starts explaining that it's
listening and what port it's listening on. This is helpful to avoid
errors if you start the blog and refresh the page before the server
has started listening on the port.

Add a basic homepage rendering test that should help protect against
regressions.

Change-Id: Ic86c30652714230b730cd4a254e69c10d019246b
Reviewed-on: https://go-review.googlesource.com/49292
Reviewed-by: Chris Broadfoot <cbro@golang.org>