Fix package name for generated code.
Following along with the article, the generated
package name in my code is the same as the original file
but this is not reflected in the article.

I am using a newer version of Go (1.8 vs 1.4) and lots
of things are different in the generated code. I feel like
those changes can be easily digested by the reader with that
level of proficiency. However, the different package name can
be confusing to new developers trying to understand how the
package and file structure works.