blog: remove all remaining protocol-relative URLs
Protocol-relative URLs (e.g., "//blog.golang.org") were previously
removed in the content of the blog, but not in other areas.
Notably, in the atom feed. This causes feed reader apps to fail, because
it doesn't know to use https for those URLs.

Always use the https scheme, including in the atom feed as well as a
couple remaining places in the blog template.

Fixes golang/go#17961

Change-Id: I694c888de9437937ba1910227ddae42d3eb0405c
Reviewed-on: https://go-review.googlesource.com/c/155357
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>