content: fix or remove broken links in Google I/O 2010 article
The original destination of these links seems to have vanished.
It doesn't really matter, so just remove them.  Leave a link to the
main Google I/O blog post, and to the video of the Go programming talk.

This particular blog article is not too interesting today but I guess
it's a reminder of the early days.

Change-Id: I8d90c84d9eb442ee4028a875d2df670f444be417
Reviewed-on: https://go-review.googlesource.com/c/blog/+/190717
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Shulhan <m.shulhan@gmail.com>