content: fix code block in Publishing Go Modules article
When indentation is used to create a code block immediately following
a list, Markdown syntax is ambiguous: it can be interpreted as another
text paragraph belonging to the last list entry, or as a standalone
code block.

The original behavior in CL 197637 made this unambiguously a code block
because the present format did not support lists. This changed when the
article was converted to use Markdown syntax in CL 222848.

Start using a fenced code block instead of indentation to resolve the
Markdown ambiguity and restore original behavior.

For golang/go#33955.

Change-Id: I9537bc72d7a32f90592942591b2688fb2f26c452
Reviewed-on: https://go-review.googlesource.com/c/blog/+/245017
Reviewed-by: Jean de Klerk <deklerk@google.com>