x/blog: add non-pointer note for errors.As example
Since this is a very high SEO article and it's a copy-pasteable block, it gets
copy pasted without much thought, which leads to problems when the error is a
concrete type / interface / etc.

This CL adds a small note that users should watch out for the case when the
error type is not a pointer.

Fixes golang/go#35361

Change-Id: Ibbd950c2a73a5f30cdab3517e042f69465adff97
Reviewed-on: https://go-review.googlesource.com/c/blog/+/252877
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Kevin Burke <kev@inburke.com>
Trust: Jean de Klerk <deklerk@google.com>
Trust: Damien Neil <dneil@google.com>