module-compatibility: make the notgrpc example use the value as previous
In the grpc.Dial option, the MaxDelay is set to time.Second, but in the
next example for a struct option the MaxDelay is set to time.Minute.

This changes make both examples use the same values for MaxDelay.

Change-Id: I0413b3ab88a1bcdc049827352b3662b98c88039d
Reviewed-on: https://go-review.googlesource.com/c/blog/+/277532
Reviewed-by: Jean de Klerk <deklerk@google.com>
Trust: Jean de Klerk <deklerk@google.com>
Run-TryBot: Jean de Klerk <deklerk@google.com>