content: add missing summaries
Also remove spurious time in godoc.org-redirect.article.
It was copy-pasted from pkgsite-redesign.article, which needed it
because we posted two blog posts together, and the times
in 11years.article and pkgsite-redesign.article decided the
order in which they appeared on the blog.
But in general times are unnecessary, so remove the
one in godoc.org-redirect.article to avoid future copy-pasting.

Change-Id: I43e43baf911155d8f8e5fbc77110d4ede836518e
Reviewed-on: https://go-review.googlesource.com/c/blog/+/278933
Trust: Russ Cox <rsc@golang.org>
Run-TryBot: Russ Cox <rsc@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>